backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
decorator==5.2.1
distro==1.9.0
Django==5.1.6
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
docstring_parser==0.16
elastic-transport==8.17.1
elasticsearch==8.17.2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.9.0
mindsdb_sdk==3.4.0
mindsdb_sql_parser==0.0.2
numpy==2.2.4
openai==1.55.3
pandas==2.2.3
psycopg2-binary==2.9.10
pydantic==2.11.0
pydantic_core==2.33.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.3
setuptools==75.8.2
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
sseclient-py==1.8.0
tenacity==9.0.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.1
urllib3==2.3.0
validators==0.20.0
